+++ examples/demo25/src/worker/worker.js
@@ -2,6 +2,10 @@ let count = 0
 
 onmessage = evt => {
     count++
-    console.log(`worker receive message(${count}): `, evt.data)
-    postMessage({from: 'worker', to: evt.data.from})
+    if (count === 1) {
+        console.log('navigator: ', navigator)
+        console.log('location: ', location)
+    }
+    console.log(`worker receive message(${count}): `, evt, evt.data)
+    postMessage({from: 'worker', to: evt.data.from, count})
 }